修改common

+++ common-service/common-bean/src/main/java/cn/linkmore/common/request/ReqAccessDetail.java
@@ -3,7 +3,7 @@ package cn.linkmore.common.request;
 public class ReqAccessDetail {
 	private Long id;
 
-    private Integer userId;
+    private Long userId;
 
     private String path;
 
@@ -27,15 +27,15 @@ public class ReqAccessDetail {
         this.id = id;
     }
 
-    public Integer getUserId() {
-        return userId;
-    }
+	public Long getUserId() {
+		return userId;
+	}
 
-    public void setUserId(Integer userId) {
-        this.userId = userId;
-    }
+	public void setUserId(Long userId) {
+		this.userId = userId;
+	}
 
-    public String getPath() {
+	public String getPath() {
         return path;
     }
